Apply ConfigureAwait(false) on all awaits
See https://msdn.microsoft.com/en-us/magazine/jj991977.aspx for
information regarding async/await. Pipeline behavior that was expecting
to run on the SynchronizationContext that the IMediator call was made
from will now have to take care of the switch back to the original
SynchronizationContext. Some examples of doing that can be found at
https://blogs.msdn.microsoft.com/pfxteam/2011/01/13/await-anything/.