Same type version for all request handlers
Initially, I got confused because I did not notice that only the ICancellableAsyncRequestHandler with a single type argument (Task return) was scanned. I wonder if it is better to either register all two arguments versions of request handlers or to drop all cancellable version of handlers as they are not used in the example. Another option could be to register all single type arguments of request handlers as well.