Fix runtime exception in DryIoc sample
The addition of the second constructor in the Mediator class (for
publishing strategies) breaks the default rules for DryIoc.

This commit will fix that by explicity picking the constructor to use in
the container. It also adds comments for an alternative solution to
automatically resolve classes that have multiple constructors.